Move the bare-minimum nodejs config to nodejs/base
Also require nodejs/file so that IO.read and friends won't break.
