+++ stdlib/opal/platform.rb
@@ -8,8 +8,7 @@ when 'gjs'
 when 'quickjs'
   require 'quickjs'
 when 'nodejs'
-  require 'nodejs/kernel'
-  require 'nodejs/io'
+  require 'nodejs/base'
 when 'headless-chrome'
   require 'headless_chrome'
 when 'opal-miniracer'
